KOOT-35: Setup project architecture
