+++ koothrappali-service/src/services/awsService.js
@@ -1,6 +1,7 @@
 import _ from 'lodash';
+import axios from 'axios';
 import AWS from 'aws-sdk';
-import CONFIG from '../const';
+import CONFIG from '../constants';
 
 AWS.config.logger = console;
 AWS.config.update({
@@ -65,38 +66,6 @@ class AwsService {
       .value();
 
     return result;
-  }
-
-  getallS3BucketNames() {
-    return new Promise((resolve, reject) => {
-      let bucketName = [];
-      let s3 = new AWS.S3({ apiVersion: '2018-10-01' });
-
-      // Call S3 to list current buckets
-      s3.listBuckets().promise()
-        .then((response) => {
-          bucketName = response.Buckets.map(a => a.Name);
-
-          return resolve(bucketName)
-        })
-        .catch(err => reject(err));
-    });
-  }
-
-  getS3BucketObjects(bucketName) {
-    return new Promise((resolve, reject) => {
-      let s3 = new AWS.S3({ apiVersion: '2018-10-01' });
-      let params = {
-        Bucket: bucketName,
-      };
-
-      // List objects in a bucket
-      s3.listObjectsV2(params).promise()
-        .then((response) => {
-          return resolve(response)
-        })
-        .catch(err => reject(err));
-    });
   }
 }
 export default AwsService;
