+++ koothrappali-service/src/services/googleOAuth.js
@@ -1,7 +1,7 @@
 import Boom from 'boom';
 import { google } from 'googleapis';
 
-import CONST from '../const';
+import CONST from '../constants';
 const { GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GOOGLE_OAUTH_CALLBACK_URL } = CONST;
 
 export const oauth2Client = new google.auth.OAuth2(GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GOOGLE_OAUTH_CALLBACK_URL);
